restructure folders around place of change
