+++ src/app/GameController.lua
@@ -1,4 +1,4 @@
-local ViewPair = require "engine.ViewPair"
+local ViewPair = require "core.ViewPair"
 
 --- Converts interaction events into meaningful object actions.
 ---@class GameController:Controller
