+++ src/app/GameRules.lua
@@ -1,6 +1,6 @@
-local GameObjectCollection = require "engine.GameObjectCollection"
+local GameObjectCollection = require "core.GameObjectCollection"
 -- [Require all game objects for game world]
-local PlaceholderObject = require "objects.PlaceholderObject"
+local PlaceholderObject = require "entities.PlaceholderObject"
 
 --- Creates and updates game objects according to game rules (not keeping the state).
 ---@class GameRules:Ruleset
