+++ src/core/GameObject.lua
